refactor(migrate): extract shared progress-notify-throttle logic
The lastNotifyAt/PROGRESS_NOTIFY_INTERVAL_MS/notificationBuilder throttle
pattern was duplicated near-verbatim between MigrationJob and
QuickMigrateJob. Extract it into MigrationProgressNotifier so future
fixes to the throttling/notification behavior only need to happen once.
